+++ src/lib/stories/actions/getStory.ts
@@ -1,0 +1,7 @@
+"use server";
+
+import * as stories from "../../repos/stories";
+
+export async function getStory(storyId: string) {
+  return stories.getStory({ storyId });
+}
